[feat] tableview에 저장된 coredata 띄우기
